+++ UnitTests/BraintreePayPalNativeCheckoutTests/BTPayPalNativeCheckoutRequest_Tests.swift
@@ -86,6 +86,7 @@ class BTPayPalNativeCheckoutRequest_Tests: XCTestCase {
         request.currencyCode = "currency-code"
         request.requestBillingAgreement = true
         request.billingAgreementDescription = "description"
+        request.userAction = .payNow
 
         let shippingAddress = BTPostalAddress()
         shippingAddress.streetAddress = "123 Main"
@@ -119,6 +120,9 @@ class BTPayPalNativeCheckoutRequest_Tests: XCTestCase {
         }
 
         XCTAssertEqual(billingAgreementDetails["description"], "description")
+
+        guard let experienceProfile = parameters["experience_profile"] as? [String: Any] else { XCTFail(); return }
+        XCTAssertEqual(experienceProfile["user_action"] as? String, "commit")
     }
 
     func testVaultParameters_withShippingAddressOverrideNil_doesNotPassShippingAddress() {